+++ src/main/java/com/example/baseballprediction/global/security/MemberDetails.java
@@ -1,0 +1,70 @@
+package com.example.baseballprediction.global.security.oauth;
+
+import java.util.Collection;
+import java.util.Map;
+
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+
+import com.example.baseballprediction.domain.member.entity.Member;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class MemberDetails implements UserDetails, OAuth2User {
+	private final Member member;
+
+	private Map<String, Object> attributes;
+
+	@Override
+	public Collection<? extends GrantedAuthority> getAuthorities() {
+		return null;
+	}
+
+	@Override
+	public String getPassword() {
+		return member.getPassword();
+	}
+
+	@Override
+	public String getUsername() {
+		return member.getUsername();
+	}
+
+	@Override
+	public String getName() {
+		return null;
+	}
+
+	@Override
+	public boolean isAccountNonExpired() {
+		return false;
+	}
+
+	@Override
+	public boolean isAccountNonLocked() {
+		return false;
+	}
+
+	@Override
+	public boolean isCredentialsNonExpired() {
+		return false;
+	}
+
+	@Override
+	public boolean isEnabled() {
+		return false;
+	}
+
+	@Override
+	public <A> A getAttribute(String name) {
+		return OAuth2User.super.getAttribute(name);
+	}
+
+	public boolean isNewMember() {
+		return member.isNewMember();
+	}
+}
